+++ packages/ripple-binary-codec/HISTORY.md
@@ -8,6 +8,10 @@
 - Exported `TRANSACTION_TYPES` value
 ### Fixed
 - Adds missing fields from XLS-20 NFT implementation
+- Fixed error being raised when decoding issued currencies in non-standard formats that decode to XRP.
+
+## 1.2.3 (2022-2-2)
+- Fix issue where ISO is invalid when parsing currency code
 
 ## 1.2.2 (2021-12-02)
 - Fix issue where unsupported currency codes weren't being correctly processed